+++ src/main/java/com/nhnacademy/store99/bookstore/auth/service/impl/AdminCheckServiceImpl.java
@@ -2,7 +2,7 @@ package com.nhnacademy.store99.bookstore.auth.service.impl;
 
 import com.nhnacademy.store99.bookstore.auth.repository.AuthRepository;
 import com.nhnacademy.store99.bookstore.auth.service.AdminCheckService;
-import com.nhnacademy.store99.bookstore.user.excepiton.UserNotFoundException;
+import com.nhnacademy.store99.bookstore.user.exception.UserNotFoundException;
 import com.nhnacademy.store99.bookstore.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
