--- src/main/java/com/nhnacademy/store99/bookstore/user/exception/UserAuthInfoNotFoundException.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.store99.bookstore.user.exception;
-
-public class UserAuthInfoNotFoundException extends RuntimeException {
-    public UserAuthInfoNotFoundException(String message) {
-        super(message);
-    }
-}
