+++ src/main/java/com/nhnacademy/store99/bookstore/user/exception/UserNotFoundByEmailException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.store99.bookstore.user.exception;
+
+import com.nhnacademy.store99.bookstore.common.exception.NotFoundException;
+
+public class UserNotFoundByEmailException extends NotFoundException {
+    public UserNotFoundByEmailException(String email) {
+        super(String.format("User not found (UserEmail: %s)", email));
+    }
+}
