+++ src/main/java/com/nhnacademy/store99/bookstore/user/service/UserService.java
@@ -3,7 +3,7 @@ package com.nhnacademy.store99.bookstore.user.service;
 import com.nhnacademy.store99.bookstore.user.dto.AuthorizationRequest;
 import com.nhnacademy.store99.bookstore.user.dto.AuthorizationResponse;
 import com.nhnacademy.store99.bookstore.user.dto.UserAuthInfoByEmail;
-import com.nhnacademy.store99.bookstore.user.exception.UserAuthInfoNotFoundException;
+import com.nhnacademy.store99.bookstore.user.exception.UserNotFoundByEmailException;
 import com.nhnacademy.store99.bookstore.user.repository.UserRepository;
 import java.time.LocalDateTime;
 import java.util.Optional;
@@ -36,7 +36,7 @@ public class UserService {
         Optional<UserAuthInfoByEmail> userAuthInfoByEmailOpt = userRepository.getUserAuthInfoByEmail(email);
 
         if (userAuthInfoByEmailOpt.isEmpty()) {
-            throw new UserAuthInfoNotFoundException("email 로 user 정보 조회 실패");
+            throw new UserNotFoundByEmailException(email);
         }
 
         UserAuthInfoByEmail userAuthInfoByEmail = userAuthInfoByEmailOpt.get();
